pyproject.toml
[project]
name = "arkitekt-server"
version = "2.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "cryptography>=45.0.5",
    "inquirer>=3.4.0",
    "namegenerator>=1.0.6",
    "platformdirs>=4.5.1",
    "py-machineid>=1.0.0",
    "pydantic>=2.11.7",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "rich-click>=1.8.9",
    "typer>=0.16.0",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

keywords = ["arkitekt", "server"]



[tool.hatch.build.targets.sdist]
include = ["arkitekt_server"]

[tool.hatch.build.targets.wheel]
include = ["arkitekt_server"]

[tool.uv]
dev-dependencies = [
    "arkitekt-next[all]>=0.17.0",
    "dokker>=2.2",
    "pytest-cov>=6.2.1",
    "pytest>=8.4.1",
    "python-semantic-release>=9.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
upload_to_pypi = true
branch = "main"
build_command = "uv build"



[project.scripts]
arkitekt-server = "arkitekt_server.main:main"
